Add function thats fits an AR model. Also fixed data download so that it is downloaded only once.

+++ src/lennart_epp/data_management/download_apple.py
@@ -8,9 +8,8 @@ from lennart_epp.config import SRC
 def download_apple_data():
     raw_data_path = SRC / "data" / "apple_data.csv"
 
-    # Ensure the directory exists
-    Path(raw_data_path.parent).mkdir(parents=True, exist_ok=True)
-
-    apple_data = yf.download("AAPL", start="2013-01-01", end="2023-12-31")
-
-    apple_data.to_csv(raw_data_path)
+    # ✅ Nur herunterladen, wenn die Datei noch nicht existiert
+    if not raw_data_path.exists():
+        Path(raw_data_path.parent).mkdir(parents=True, exist_ok=True)
+        apple_data = yf.download("AAPL", start="2013-01-01", end="2023-12-31")
+        apple_data.to_csv(raw_data_path)
